CI: add runtime tests in qemu
This adds a github action which will build the MIPS malta 32 BE and the
armvirt 64 target and then starts it in qemu. In qemu we execute some
automated tests to check if basic functionality is working.

The script which starts and controls the qemu system is written in
python and intentionally not directly coupled with github, this allows
to run it also with other CI systems. The qenum script checks if
changing variables in uci work, that https client works and that we can
do an ssh login to the system. This should get extended in the future.

The github action is split into two parts by intention, this allows us
to later build the image in the cloud and then run it on an own github
action runner which can interact with real hardware.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
[use external repository with tests]
Signed-off-by: Paul Spooren <mail@aparcar.org>