buildsystem: add PKG_CVE_FIXED and PKG_CVE_IGNORE
In a stable branch it may not be possible to update the version to close
a CVE. However, a patch can be applied that fixes exactly this CVE. The
problem is that CVE scanners don't know about it because they only look
at the version.

For this reason two no PKG Variables could be added to the packages
makefile.

PKG_CVE_FIXED:
Variable for defining CVEs that do apply, but aren't properly marked as
fixed at cve.org in the current version.

PKG_CVE_IGNORE:
Variable for defining CVEs that don't apply due to features not enabled,
or affecting other platforms (e.g Windows).

Signed-off-by: Florian Eckert <fe@dev.tdt.de>