tools/coreutils: use automake during bootstrap
This was previously not possible because
part of the source code for coreutils in the form
of extra gnulib modules was missing from the release distribution.

Now that the local modules from coreutils source
is included in coreutils releases,
the bootstrap script can be ran without skipping
automake in the autoreconf stage after importing modules
by using fake paths to each autotools program,
and instead use the real paths to our modified autotools.

Because we do not require tool for building documentation,
continue to fake the paths to autopoint and gtkdocize.

Remove skipping the import of some source files
which is no longer necessary.

Signed-off-by: Michael Pratt <mcpratt@pm.me>