base-files: sysupgrade: use /etc/apk/world for -k
With the introduction of apk, we now have /etc/apk/world which is a list
of packages selected to be installed (excluding automatic dependencies).
For reinstalling a new system, that is normally better than the raw list
of packages.

We can compare /rom/etc/apk/world and /etc/apk/world to detect the
origin of each installed packages. If any of those files are missing,
the old method is used.

This new check also introduced the package state "removed" for packages
present in the rom but removed from the overlay.

Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>