ipq40xx: Add support for Meraki Go GX20
Wired router with 5 port Gigabit switch. It shares the same PCB as the Z3, but without the WiFi radios.

Port 5 has POE output (802.3af). The WAN port is used for tftp booting in U-Boot.

This device ships with secure boot, and cannot be flashed without an external programmer.
More details on secure boot can be found at:
https://watchmysys.com/blog/2024/04/breaking-secure-boot-on-the-meraki-z3-and-meraki-go-gx20/

Information about flashing can be found on the OpenWrt Z3 device page.

Signed-off-by: Hal Martin <1230969+halmartin@users.noreply.github.com>