ipq40xx: Add support for Meraki MR30H
PoE powered 802.11ac access point with an integrated 5 port Gigabit switch.

This device is configured with only LAN ports, as there is no "WAN" port by default.

Port 4 has PoE output (802.3af) and is the port used for tftp booting in U-Boot.

The device does not have secure boot enabled at this time.

Stock U-Boot does not allow interrupting boot to obtain a shell,
so it is highly recommended to flash a custom U-Boot build.
An external programmer is required to overwrite the U-Boot region of NAND,
similar to other recent Meraki devices such as the MR33.

Signed-off-by: Hal Martin <1230969+halmartin@users.noreply.github.com>