WIP: realtek: implement polling for hardware counters
123456789012345678901234567890123456789012345678901234567890123456789012

Some stats are calculated from multiple hardware counters, some of which
are 64 bits and some are 32 bits. In the case where one of the 32 bit
counters overflows, this can produce weird results.

To fix this, maintain 64 bit counters by polling the hardware counters
and adding up the differences. Polling needs to happen just often enough
to catch every single overflow.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>