ath79: iodata: set ath9k LED pin
The ath9k driver creates an ath9k LED by default. Instead of having a
non functional LED, configure it properly and remove the extra as it's
not needed.

Set the pcie device to qcom,ath10k as the Documentation says, The pci
line implies it's ath9k when it's actually ath10k.

Signed-off-by: Rosen Penev <rosenp@gmail.com>