ath10k-ct-firmware: update wave2 chip firmware to latest beta version
Release Notes:

* Support enabling/disabling rts/cts. Not sure if this really works as
  hoped though.

* Improved compiler options to build smaller sized binaries.

* Sept 15, 2022:  Fix per-peer txpower to not over-ride tpc when value
  is 0 (ie, not set yet). This in turn fixes TPC over-ride through the
  txo path.

* April 4, 2023:  Fix uninitialized minccapwr threshold variables,
  reported by Sebastian.

Candela Technologies has already used this beta firmware in their
commercial product "LANforge". And the -ct firmware homepage states
that "This is code is not being changed much, and should be considered
for stable use now." On QCA9888 and IPQ4019, I didn't catch any speed
or stability issues compared to the previous version. Therefore, I
think it is worth updating this beta version.

Signed-off-by: Shiji Yang <yangshiji66@outlook.com>